dbt_source_codes/temp_spark/requirements.txt
functions-framework
google-cloud-dataproc